backend/pyproject.toml
[project]
name = "my-rag-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0,<2.8",  
    "streamlit>=1.30.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]